fix: mypage, editpage 분리
