+++ package.json
@@ -5,22 +5,23 @@
   "scripts": {
     "dev": "next dev -p 4000",
     "build": "next build && next export",
+    "start": "next start",
     "lint": "next lint --fix",
     "postinstall": "husky install"
   },
   "dependencies": {
     "@emotion/react": "^11.11.0",
     "@emotion/styled": "^11.11.0",
-    "@next/bundle-analyzer": "^13.4.6",
+    "@next/bundle-analyzer": "^13.5.4",
     "@tanstack/react-query": "^4.29.5",
     "@tanstack/react-query-devtools": "^4.29.6",
-    "@team-return/design-system": "^1.0.80",
-    "axios": "^1.4.0",
+    "@team-return/design-system": "^1.1.3",
+    "axios": "^1.5.1",
     "debug": "^4.3.4",
     "eslint": "8.38.0",
-    "eslint-config-next": "13.3.0",
+    "eslint-config-next": "^13.5.4",
     "framer-motion": "^10.12.10",
-    "next": "^13.4.4",
+    "next": "^13.5.4",
     "next-cookie": "^2.8.0",
     "next-cookies": "^2.0.3",
     "next-optimized-images": "^2.6.2",
@@ -35,7 +36,6 @@
     "styled-components": "^6.0.0-rc.1",
     "styled-reset": "^4.4.7",
     "supports-color": "^9.3.1",
-    "typescript": "5.0.4",
     "typescript-eslint": "^0.0.1-alpha.0",
     "zustand": "^4.3.8"
   },
@@ -44,7 +44,7 @@
     "@testing-library/react": "^14.0.0",
     "@types/debug": "^4",
     "@types/node": "^18.15.11",
-    "@types/react": "^18.0.35",
+    "@types/react": "18.2.25",
     "@types/react-beautiful-dnd": "^13.1.4",
     "@types/react-cookies": "^0",
     "@types/react-dom": "^18.0.11",
@@ -64,7 +64,8 @@
     "jest-environment-jsdom": "^29.5.0",
     "postcss": "^8.4.23",
     "tailwind-styled-components": "^2.2.0",
-    "tailwindcss": "^3.3.1"
+    "tailwindcss": "^3.3.1",
+    "typescript": "^5.2.2"
   },
-  "packageManager": "yarn@3.6.0"
+  "packageManager": "yarn@3.6.4"
 }
